파일 구조화, 모듈화
